Add TcpListener class + Move tcp state machine to Tcp.cs
